started to refactor robot class
